Renames dbConfig to index.js and adjusts imports
